[Update] Review of pull request
