+++ include/Robot.h
@@ -5,15 +5,15 @@
 
 class Robot
 {
-private:
-    Color color;
-    std::pair<int, int> position;
-public:
-    Robot(Color c, std::pair<int, int> p);
-    Robot(Color c, int x, int y);
-    Color getColor();
-    std::pair<int, int> getPosition();
-    int getX();
-    int getY();
-    void move(int x, int y);
+    private:
+        Color color;
+        std::pair<int, int> position;
+    public:
+        Robot(Color c, std::pair<int, int> p);
+        Robot(Color c, int x, int y);
+        Color getColor();
+        std::pair<int, int> getPosition();
+        int getX();
+        int getY();
+        void move(int x, int y);
 };
